updated uninstall script

+++ uninstall.php
@@ -6,11 +6,11 @@
  * @package NRDUntappdImporter
  */
 
-if (!defined('WP_UNINSTALL_PLUGIN')) {
+if (defined('WP_UNINSTALL_PLUGIN')) {
 
   $custom_post_types = get_option('nrd_untappd_importer_menu') ?: array();
 
-  if(count($custom_post_types) > 0) {
+  if (count($custom_post_types) > 0) {
     // Delete all posts
     foreach ($custom_post_types as $post_type) {
       $posts = get_posts([
@@ -20,26 +20,28 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
       ]);
 
       foreach ($posts as $post) {
-        wp_delete_posts($post->ID, true);
+        wp_delete_post($post->ID, true);
       }
-    }
 
-    // Delete all terms of associated taxonomies
-    $taxonomies = get_object_taxonomies($post_type);
-    foreach ($taxonomies as $taxonomy) {
-      $terms = get_terms(
-        array(
-          'taxonomy' => $taxonomy,
-          'hide_empty' => false,
-        )
-      );
+      // Delete all terms of associated taxonomies
+      $taxonomies = get_object_taxonomies($post_type['post_type']);
+      foreach ($taxonomies as $taxonomy) {
+        $terms = get_terms(
+          array(
+            'taxonomy' => $taxonomy,
+            'hide_empty' => false,
+          )
+        );
 
-      foreach ($terms as $term) {
-        wp_delete_term($term->term_id, $taxonomy);
+        if (!is_wp_error($terms)) {
+          foreach ($terms as $term) {
+            wp_delete_term($term->term_id, $taxonomy);
+          }
+        }
       }
     }
   }
-  
+
   $options = [
     'nrd_untappd_importer_schedule',
     'nrd_untappd_importer_api_creds',
@@ -52,5 +54,3 @@ if (!defined('WP_UNINSTALL_PLUGIN')) {
     delete_option($option);
   }
 }
-
-// TODO: Clear Database stored data
